+++ src/pages/Dashboard/Dashboard.styles.ts
@@ -2,8 +2,4 @@ import styled from "styled-components";
 
 export const Container = styled.div`
   padding: 24px;
-  display: flex;
-  flex-direction: column;
 `;
-
-
