+++ src/pages/Dashboard/components/Columns/Columns.styles.tsx
@@ -19,11 +19,11 @@ const registrationStatusStyles: {
 };
 
 export const Container = styled.div`
-  display: grid;
-  grid-template-columns: 1fr 1fr 1fr;
-  grid-gap: 24px;
-  justify-content: center;
+  display: flex;
+  gap: 24px;
   margin-top: 24px;
+  width: auto;
+  overflow: auto;
 `;
 
 export const Column = styled.div<{ status: RegistrationStatus }>`
@@ -31,8 +31,10 @@ export const Column = styled.div<{ status: RegistrationStatus }>`
   background-color: ${({ status }) =>
     registrationStatusStyles[status].background};
   border-radius: 32px;
-  min-height: 80vh;
-  max-height: 80vh;
+  height: 80vh;
+  flex-shrink: 0;
+  min-width: 400px;
+  flex: 1;
 `;
 
 export const TitleColumn = styled.h3<{ status: RegistrationStatus }>`
